Moved PanelList to separete file
