+++ src/components/StudentPage/RepositoriesList.tsx
@@ -1,9 +1,7 @@
 import { useState, useEffect } from 'react';
 import { Collapse,Button,Space } from 'antd';
-import { StudentRepository } from './interfaces';
 import { Urls } from './constants';
-
-const { Panel } = Collapse;
+import {PanelList} from './PanelList';
 
 export const RepositoriesList = (prop: { user: string }) => {
   //TODO: Перенести в эффектор. Сделать связку со store User,
@@ -30,24 +28,12 @@ export const RepositoriesList = (prop: { user: string }) => {
         (setAllRepoShown(true), setRepositoriesPerPage(100))
     )
   }
-
-  const panelList = repositories.map((item: StudentRepository) => {
-    return (
-        <Panel header={item?.name} key={`${item.id}`}>
-          <p>URL репозитория:</p>
-          <p>
-            <a href={item.html_url ?? ''}>{item.full_name ?? ''}</a>
-          </p>
-          <p>Описание: {item.description|| "осутствует"}</p>
-          <p>Дата создания: {(item.created_at).slice(0,10).split("-").reverse().join(".")}</p>
-        </Panel>
-    )
-  })
+console.log('репо', repositories);
 
   return (
     <>
       <Collapse defaultActiveKey={[]} >
-        {panelList}
+        <PanelList repositories={repositories}/>
       </Collapse>
       <Button type="default" onClick={handleClick}>
         {allRepoShown ? "Скрыть " : "Показать "}
