+++ src/components/StudentPage/interfaces.ts
